# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product API",
        "description": "A simple CRUD API for managing products",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get all employees",
                "operationId": "326f2e86fa40a9ba88b1cb5f2d4b62b8",
                "responses": {
                    "200": {
                        "description": "List of employees"
                    }
                }
            },
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Create a new employee",
                "operationId": "4a10b3ada27736863f2586c56b9c19ec",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Employee created"
                    }
                }
            }
        },
        "/api/v1/employees/{id}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get a single employee",
                "operationId": "b7cb6100cedcac60d8c23d3f9d7c3251",
                "responses": {
                    "200": {
                        "description": "Employee found"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Employees"
                ],
                "summary": "Update an employee",
                "operationId": "bcdffcad24dd066f489c0e6ed83861a5",
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Employees"
                ],
                "summary": "Delete an employee",
                "operationId": "43aa0966dc246d0b0373a7db65a6bf78",
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    }
                }
            }
        },
        "/api/v1/products": {
            "get": {
                "summary": "Get all products",
                "operationId": "1ce27f90dd8c39cea9520bf45d58c7c2",
                "responses": {
                    "200": {
                        "description": "List of products"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Employees"
        }
    ]
}